Update measureContigLengthFromFasta.sh
rather than export the output to standard out, the user now must set it to an output file path in the parameters. In addition, I added an additional parameter for users to set whether they want to use more RAM or more long-term memory. I also made it so that if the fasta file is decompressed, the code with compress it back afterwards.